+++ src/mentor/infrastructure/entity/mentor.entity.ts
@@ -1,9 +1,9 @@
 import { Column, Entity, JoinColumn, OneToMany, OneToOne } from 'typeorm';
-import { BaseEntity } from '../../common/entity/base.entity';
-import { User } from '../../user/infrastructure/entity/user.entity';
+import { BaseEntity } from '../../../common/entity/base.entity';
+import { User } from '../../../user/infrastructure/entity/user.entity';
 import { MentorTechStack } from './mentor-tech-stack.entity';
 import { MentorInterest } from './mentor-interest.entity';
-import { MentorAccept, MentoringType } from '../enum/mentor.enum';
+import { MentorAccept, MentoringType } from '../../enum/mentor.enum';
 
 @Entity()
 export class Mentor extends BaseEntity {
