+++ src/mentor/mentor.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Mentor } from './entity/mentor.entity';
-import { MentorTechStack } from './entity/mentor-tech-stack.entity';
-import { MentorInterest } from './entity/mentor-interest.entity';
-import { MentorController } from './mentor.controller';
-import { MentorService } from './mentor.service';
+import { Mentor } from './infrastructure/entity/mentor.entity';
+import { MentorTechStack } from './infrastructure/entity/mentor-tech-stack.entity';
+import { MentorInterest } from './infrastructure/entity/mentor-interest.entity';
+import { MentorController } from './presentation/mentor.controller';
+import { MentorService } from './application/mentor.service';
 import { InterestModule } from '../interest/interest.module';
 import { TechStackModule } from '../tech-stack/tech-stack.module';
 import { UserModule } from '../user/user.module';
