+++ src/mentor/presentation/res.dto.ts
@@ -6,7 +6,7 @@ import {
   IsOptional,
   IsString,
 } from 'class-validator';
-import { Mentor } from '../entity/mentor.entity';
+import { Mentor } from '../infrastructure/entity/mentor.entity';
 import { User } from '../../user/infrastructure/entity/user.entity';
 import { MentorProfileReqDto } from './req.dto';
 import { MentorAccept, MentoringType } from '../enum/mentor.enum';
